+++ Runtime/RegionsDataPropertyAttribute.cs
@@ -1,8 +1,0 @@
-using UnityEngine;
-
-namespace ActionCode.Cinemachine
-{
-    public sealed class RegionsDataPropertyAttribute : PropertyAttribute
-    {
-    }
-}