+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Begin region transition event
 
+### Removed
+- Redundant RegionsDataPropertyAttribute
+
 ### Changed
 - Fix not saving Regions Data on ScripableObject